requirements.txt
playwright>=1.44,<2.0
python-dotenv>=1.0.1
